pyproject.toml
[project]
name = "i18n-readme"
version = "0.1.0"
description = "auto generate multilingual README.md"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-adk>=1.5.0",
    "litellm>=1.73.6.post1",
]

[project.scripts]
readme-trans = "i18n_readme.cli:main"

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
]
